Updates for reading ntuples provided by Nick

+++ interface/miniTree.h
@@ -47,7 +47,7 @@ class miniTree {
     configuration * m_config;
 
     // new branches defined here 
-    double m_dnn;
+    float m_dnn;
 };
 
 #endif
